requirements.txt
telethon
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver
python-dotenv
psycopg2-binary # For PostgreSQL connectivity